feat: start point for LSP API upgrade
require("lspconfig") set to be deprecated. Have put an example config
that will work for css in the init, but should more broadly replace the
examples from lua/plugins/lspconfig.

also packaged in some config files for either what we use now or are
keen to experiment with in the future from
https://github.com/neovim/nvim-lspconfig/tree/master/lsp